Finishing Detail

+++ app/src/main/java/com/catnip/mycoin/data/network/model/response/coin/detail/CurrentPrice.kt
@@ -4,5 +4,5 @@ import com.google.gson.annotations.SerializedName
 
 data class CurrentPrice(
     @SerializedName("usd")
-    val usd: Int?
+    val usd: Double?
 )
